feat: unlock lessons after quiz completion
